Fix java version check if java is unavailable (#1730)
## Changes

### What does this PR do?

Prior to installing the morpheus transpiler we verify the version of
Java that is available. The current code fails if Java is not installed
at all; this PR updates the check so that it properly detects the
complete absence of Java.

### Relevant implementation details

The `CalledProcessError` exception is only raised if the process could
not be _started_. This does not include a failure to find the executable
to start.

### Caveats/things to watch out for when reviewing:

This PR scoped as narrowly as possible for dealing with the bug.

### Linked issues

Resolves #1728.

### Functionality

- modified existing command: `databricks labs lakebridge
install-transpile`

### Tests

- manually tested
- added unit tests

### Demo

After blocking `/usr/bin/java` the logs show this during
`install-transpile`:
```
13:03:21    DEBUG [d.l.lakebridge.install] Initializing workspace installation for module: transpile (config: None)
13:03:21     INFO [d.l.lakebridge.install] databricks-bb-plugin v0.1.6 already installed
13:03:21  WARNING [d.l.lakebridge.install] This software requires Java 11 or above. Please install Java and re-run 'install-transpile'.
13:03:21     INFO [d.l.lakebridge.install] Configuring lakebridge `transpile`.
```