Refactor parsing of `java -version` during `install-transpile` (#1731)
## Changes

### What does this PR do?

This PR replaces the custom parsing we had in place for the installed
version of Java. Changes include:

 - Handling the version as a 4-tuple.
 - Using a regex instead of a custom parser.
- The [specification][1] allows a version to be just a single digit (eg.
`28`); this would cause the previous code to crash.

Further to this, to aid with support:

 - We now log the version of java that was detected.
- If an error occurs while running `java -version` we now log
information about what happened.

### Relevant implementation details

The specification for java version strings is [here][1].

### Caveats/things to watch out for when reviewing:

Optional groups within a pattern yield `None` if they weren't part of
the match.

### Linked issues

This follows on from [a
comment](https://github.com/databrickslabs/lakebridge/pull/1537/files#r2075491051)
on #1537, where this code was originally implemented.

### Functionality

- modified existing command: `databricks labs lakebridge
install-transpile`

### Tests

- manually tested:
  ```
15:20:20 DEBUG [d.l.lakebridge.install] Detected java version: (24, 0,
1, 0)
  ```
- added unit tests
- existing integration tests

[1]:
https://docs.oracle.com/en/java/javase/11/install/version-string-format.html